requirements.txt
asn1crypto==1.5.1
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
oscrypto==1.3.0
pdf2image==1.17.0
pillow==11.3.0
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyMySQL==1.1.1
pypdf==6.0.0
PyPDF2==3.0.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.2
requests==2.32.4
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
Werkzeug==3.1.3
